MoreChanges

+++ DataLayer/Entities/OrderDetails.cs
@@ -14,7 +14,7 @@ namespace DataLayer.Entities
 
         public string ItemName { get; set; }
 
-        public decimal ItemQty { get; set; }
+        public int ItemQty { get; set; }
 
         public decimal ItemTotalPrice { get; set; }
     }
